Make it clear what `kritis` is
One of the examples demonstrates building an arbitrary pipeline for
a project called `kritis`, mostly b/c folks who are working on this
project have also worked on `kritis` and are familiar with it, but it's
not actually relevant to `bulid-pipeline` directly. Tried to make this
more clear in the docs.

Also in #2 we had a question about what `strawman` means (not the first
time this has come up!) A defintion is avialable at https://en.wikipedia.org/wiki/Straw_man_proposal
but also mentions it is "american business jargon" so if we want to go
on using that term we should define it. Instead I removed it from the
one place it exists in our docs. Since no one has completely beaten up
the proposal yet it seems to be graduating beyond the "strawman" phase -
but correct me if I'm wrong :D