Add CONTRIBUTING.md and code-of-conduct.md
The contributing.md includes info on how pull requests work
(e.g. prow commands post #41), principles around including
tests, docs, and commit messages.

It also includes the current state of the roadmap and how
to find issues we'd like contributions on.

The code of conduct is the same code of conduct used by the
rest of the knative projects.