Add "valid" condition to all resources
All resources will be evaluated by their controllers after they are
created to determine if they are valid or not. This PR adds a "Valid"
condition to the status of all resoruces which can be used by the
controller to inform the user of this (otherwise I think we'd have to
implement this kind of validation via admission controllers? maybe
that's better?)

This would be used by the logic that will be implemented for validation
in #33 #32 #31 #29 #30.