Add issue number to commit message requirements
Not every commit message will have an issue number associated with it,
(this one doesn't! hah!) but if it does, it should reference that issue
number somewhere in the body, and if the commit resolves the issue,
take advantage of GithHub's automation to automatically link the two
and resolve the issue.