Kick off unit tests from prow
This uses the `report_go_test` function which is responsible for
producing unit tests in a format that can be understood by testgrid. We
aren't using test grid yet but I guess why not get ready for it.

At it's core the function calls `"go test -race -v ${args[@]}"` which
has the added bonus of detecting race conditions so why not I guess.
It doesn't do anything with coverage yet tho so we'll have to tackle
that later.

Fixes #78