Create namespace for tests to use
Before running any tests, create a randomized namespace, and tear down
the namespace at the end of the test. Because we're using `os.Exit`, we
can't use `defer` to execute the cleanup so we'll just do it the old
fashioned way.

Also added an empty `tearDown` to the integration test to demonstrate
how it should be added (specifically to make sure to run it on
interrupt as well).