Make `namespace` private to `test` package
I made it public (`Namespace`) because of briefly misunderstanding the
scope of private variables (I thought they were limited to the file, but
they are available to the whole package).

Also changed some methods to use the global variable, this seems
generally worse but on the other hand there's no reason at the moment
for these methods to operate on any namespace other than the global one.