Add skeleton integration test (#74)
* Add skeleton integration test

In this first step, we add an empty integration test that does nothing
but list `Pipeline`s in a namespace harded coded to `arendelle`
(already used the Toy Story theme in knative/serving, needed something
different here, thanks @aaron-prindle XD).

This can be run manually against a developer's own cluster, so folks can
start adding integration tests in parallel if they want to.

This is based off the end to end tests in knative/serving
(https://github.com/knative/serving/tree/master/test#running-end-to-end-tests)
and uses the libs in knative/pkg extensively, which were made from the
libs in knative/serving. The docs weren't moved with the libs tho which
I'm trying to fix in:
* https://github.com/knative/serving/pull/2097
* https://github.com/knative/pkg/pull/102

The commit adds opencensus, since this is used by the test libs for
collecting and emitting metrics. We might want to use those eventually
too so that seems fine tho it would be nice if they were optional
¯\_(ツ)_/¯

This is the first step for #16, next I'll be looking into executing these
via Prow, either against a hardcoded cluster with the secrets in Prow or
going straight for kubetest + boskos.

* Create namespace for tests to use

Before running any tests, create a randomized namespace, and tear down
the namespace at the end of the test. Because we're using `os.Exit`, we
can't use `defer` to execute the cleanup so we'll just do it the old
fashioned way.

Also added an empty `tearDown` to the integration test to demonstrate
how it should be added (specifically to make sure to run it on
interrupt as well).

* Make `namespace` private to `test` package

I made it public (`Namespace`) because of briefly misunderstanding the
scope of private variables (I thought they were limited to the file, but
they are available to the whole package).

Also changed some methods to use the global variable, this seems
generally worse but on the other hand there's no reason at the moment
for these methods to operate on any namespace other than the global one.