Run integration tests on presubmit
Using the scripts from knative/test-infra, run the integration tests
against a cluster created with boskos. A lot of the boskos stuff is
still kind of magical, for example I have no idea what is controlling
the $PROJECT_ID that is being used by knative projects on presubmit, but
after wrestling with it a bit and hacking around `--run-tests` I was
able to get it to work locally and deploy a cluster (Using boskos if
running from Prow, or kubetest otherwise). I'll follow up on this more
later on and update the docs when I have more info.

Fixes #16