Add PipelineRun integration test and refactor unit tests
Added the beginnings of a skeleton of an integration test for
PipelineRun for #61. In doing this I realized that I couldn't create
PipelineParam objects with the clients b/c the plural name was wrong -
it needs to be `pipelineparamses` like I'm Smeagol. I tried to work
around this and stumbled on
https://github.com/kubernetes/code-generator/issues/53 so it doesn't
seem to be possible without changing the code-generator code :(

Meanwhile refactored the existing controller code to break some code out
of the Reconciler, so it can be instantiated without an entire
controller (instead depends only on the objects it needs). The decision about
what TaskRuns to create has been separated from the logic to retrieve
existing ones, and the logic to create them. (`GetTasks`,
`getNextPipelineRunTaskRun`)

The tests were refactored such that the success cases are in separate
tests from the failure cases, so the table driven tests dont have to
handle both and it's more clear what the tests are doing.