Remove fields from examples which aren't in spec
None of these fields are in our current spec, so let's remove them for
now.

As far as `nextTask` and `prevTask` are concerned, we're still going to
need some way to represent the order the Tasks actually executed in
while the Pipeline was running (maybe? or would we infer this from the
Pipeline?) but we can remove them for now.