Move `GetNextPipelineRunTaskRun` to `resources` package
The convention across other knative projects is that when a reconciler
needs to access other types of resources, the logic for doing this is
put in a `resources` package in the reconciler's dir.

While doing this I also abstracted the dependency on the TaskRun lister
so that we can more easily mock it out without having to involve Fake
libraries or seed any data.

This is a bit inconsistent with the function `Pipeline.GetTasks` -
theoretically `GetNextPipelineRunTaskRun` could be an attribute on
`Pipeline` as well but since it has some extra logic around naming
`TaskRuns` I think the reconciler package maybe makes more sense for
now.