Combine logic to grab Tasks and their Runs for a PipelineRun
It turns out that we need to look at TaskRuns for a few reasons,
including 1) figuring out what to run next and 2) determining the status
of the PipelineRun, so I've refactored the logic that grabs these to
collect a bunch of related state that can be reused.

When the graph becomes more sophisticated, we will need to make this
structure more than just a list.