Add validation for PipelineParams
This PR adds vaildation for the PipelineParams type. It makes sure a
Service Accounts is specified.For Results, it
makes sure that Runs and Logs have a valid URL and Type. If
Results.Tests is present, is does the same checks for Tests. If Clusters
are present, it makes sure each Cluster has a valid Endpoint and Type.

I tried to ping the Cluster Endpoint to make sure it exists, but
apparently "ping" in Go requires root access. Research could definitely
be done to find a way to do this.