Add/cleanup end to end debug info
Something has gone wrong with one of the integration tests on my PR and
I don't know what so I'm trying to add more info.

Added Builds to the dumped CRDs, and also moved the step that deploys
the examples is now after the integration tests b/c it produces a lot of
errors in the logs (hahaha...) and makes it harder to debug integration
tests failures.