Change `TaskRef` to string instead of a struct with just one field.
Right now, the `TaskRef` struct only declares field `Name`.
Defining a struct for a just 1 field makes yaml longer.
In attempt to make the pipleine config concise, this PR remoes `TaskRef` and replaces it with a string.
In future if we decide to add other fields like APIVersion, we can write reference it as "<taskname>:<version>".
Right now APIVersion is not declared and used at all, so removing it.