Add validation for PipelineParams (#141)
* Add validation for PipelineParams

This PR adds vaildation for the PipelineParams type. It makes sure a
Service Accounts is specified.For Results, it
makes sure that Runs and Logs have a valid URL and Type. If
Results.Tests is present, is does the same checks for Tests. If Clusters
are present, it makes sure each Cluster has a valid Endpoint and Type.

I tried to ping the Cluster Endpoint to make sure it exists, but
apparently "ping" in Go requires root access. Research could definitely
be done to find a way to do this.

* Add validation to webhook and make sure PipelineParams complies with
webhook.genericCRD interface

I added the validation check to the webhook, and made sure
PipelineParams complies with webhook.genericCRD. Since
webhook.genericCRD must comply with apis.Defaultable and
apis.Validatable I didn't explicitly specify those APIs, but I did add
pipelineparams_defaults.go to comply with those requirements.

* Address code review comments

serviceAccount isn't required, so I removed checks for that. Also, only
check a URL if one is passed through. I also split up the unit tests so
that they only test one thing at a time which should be easier to
maintain.