Move controller with fake data into common lib
Instead of duplicating the logic to declare and seed test data for the
controllers, this moves the logic into a common lib that both sets of
controller tests can share.

This is in preparation for #174 which will try to make sure that the error
handling and status updating for both kinds of Runs is in sync.

I think in the long run we should improve our test lib docs, and maybe
move the lib docs out of `test/README.md` into purely go docs.