Grouped invalid Run tests together
For each of TaskRun and PipelineRun there is now an `invalid` test case
for when the Run is considered invalid (and reconciling should stop).

Also changed the log level for invalid Runs to Info b/c this means the
_user_ made an error, the Error log level should be used for when there
is an Error in the system/controller itself.

For the record I'd like to say that I disagree with asserting against
log messages; I think this makes for brittle tests and very little gain,
but I think in this instance this is the best way to be sure the
reconciler is behaving as we want (since we want it to return nil
errors). I think the keys is going to be moving more logic out of the
reconciler itself into other functions we can test separately.