Use 'reason' and 'message' fields in status
Instead of relying on log messages, in most cases we can actually return
the error to the user by including it in the `message` field! Thanks for
the enormous improvement shashwathi@!

One thing I don't like at the moment is that a pipelinerun's Reason
constants are spread between `pipelinerun` and `resources`.