Update docs about Builder image
The Builder image contract is actually the opposite of what we now
support - it specifies that users should define an `entrypoint` and rely
on that, however the new behaviour in #167 will completely ignore this
`entrypoint`.

These docs are a copy of the Builder image docs at
https://github.com/knative/docs/blob/master/build/builder-contract.md
but modified to describe the contract required by Task.

This is for #143