Add Pipeline DAG.
`reconciler.pipeline.resources.Build` creates DAG for a given Pipeline.

This DAG is represented as a List of Nodes.
Each DAG `Node` has reference to the PipelineTask in the PipelineSpec and
list of task it depends on in `Node.Prev`
For now, the DAG is one-directional only i.e. from current PipelineTask you can
get to previous `PipelineTasks` and not the other way around.
This basic implementation of DAG with no traversing.

The `Build` function raises an error if the PipelineSpec is invalid.
It validates the graph and raises error when
1. A pipeline defines multiple PipelineTask with same name in the PipelineSpec.Tasks
2. A pipeline adds a dependency on PipelineTask which does not exists.
3. A pipeline has a cycyle.
The first two validations are also handled in the Pipeline Validation during creation.
However, it does not harm to add it in here.