Use 'name' instead of 'key' for binding resources
We need to create bindings between the Resources Tasks need and the
Resources we have. We were previously using `key` to identify which of a
Task's Resource needs we were binding one of our existing Resources to,
but @pivotal-nader-ziada pointed out that `Name` is a more
straightforward and clear option.

I considered updating the cluster binding as well, but since #68 is a
work in progress I figured it would cause less conflicts to leave it.

I also deleted `PipelineResourceVersion` b/c this seems to duplicate
`TaskRunResourceVersion`, only exists in `Status`, isn't currently being
used and doesn't have a clear purpose at the moment. We can add those
fields back into `Status` if we feel like we need them later.