Run PipelineRun tests in parallel
In https://github.com/knative/build-pipeline/pull/197 @shashwathi
updated the pipelineRun tests so that they could be run as table driven
tests. By making it so that each subtest creates its own namespace and
logger, we can run the tests in parllel. They will not interfere with
each other, and log messages will be interweaved, but the logger will be
named after the tests so it will be possible to reconstruct which log
message is from which test.

One side effect of this is that the metrics gathering piece is global,
so we cannot trust metrics which are emitted from these tests. I
considered making a follow-up ticket to address this, but since we
aren't using these metrics at all currently it seemed premature.

Running the tests in parallel reduces the time these tests take to run
by ~ half, e.g. from 180 seconds to 70-100 seconds in my runs.