Add mutex around logger creation when running in parallel
Unfortunately initializing the context logger sets the global metrics
collector as well, which means that when running tests in parallel, each
with their own logger, there is a race condition.

If we use a mutex when creating the loggers we can avoid this race
condition, however this is kind of a terrible fix b/c we're still
overwriting the global metrics collector.

I'll contribute this "fix" upstream as well though the underlying
problem of the global metrics collector still exists :(