Update roadmap + instructions for assigning issues
@wtam2018 indicated that they were interested in working on an issue,
but weren't sure how to assign the issue to themselves. @vdemeester
helped out but it still wasn't clear what the process is.

This commit adds some info about the process - unfortunately we are
still limited by the groups the user is in - it would be nice if any
user could assign issues to themselves, but this will have to do for
now.

I also updated the roadmap since those docs were in the same place.

Removed the project board since we are using zenhub now.