Clean up e2e-tests.sh
Remove unnecessary code and sync up the implementation with the other scripts in the other repos (serving, building, etc).

The removal of unconditionally overwriting `DOCKER_REPO_OVERRIDE` is the most important part of this PR, as it's not only unnecessary but also the fix for #253.

This PR also updates `test-infra` to include the latest fixes and features, while also updating the `Gopkg` files so the cleanup workaround in `update-deps.sh` is not necessary anymore.