Add library for resolving TaskRuns
In the TaskRun reconciler, there are several different places where
logic exists to retrieve references Tasks and Resources; this library
will allow us to retrieve these references up front and use them instead
of using Listers everywhere.

This is a step toward fixing #213 (the goal is to reuse and simplify
this logic).