Finally commit a demo example that works :D
This example will build two images and then deploy them. It uses the
microservice example in Skaffold (just an arbitrary choice of a simple
deployment situation with 2 apps).

This is the (updated version of) the demo that @dlorenc gave at Jenkins
World.

It's kind of weird that this adds yet another example; I suggest that as
part of #108 we remove the other examples since we've never actually run
them successfully: they were added at the beginning of the project as
examples of where we wanted to go.

Had to update step names b/c using camelcase in container names is not
allowed (opened #273 to handle this).

Fixes #89