Remove guestbook and kritis examples
In #108 (where we want to add test coverage to our examples)
@pivotal-nader-ziada struggled to get the examples to work and
had a lot of trouble. This was b/c we had created these examples well
before we had anything actually working, so they were more to show off
the API than anything else. Instead of wrestling with these to make them
work, we can remove them in favour of the simpler example which we know
(at least currently) works.

In the future we can add additional complexity to this Pipeline or just
add more examples.

Also keeping the example that demonstrates passing inputs and outputs
between Pipelines - I tried it locally and it works! Just gotta get that
automation from #108 in place so we can keep it that way :D