Add `Results` fields to `TaskRun` and `PipelineRun` status
When a user kicks off a run, they will provide an endpoint to upload
logs to (initial implementation will be in #107). The corresponding
fields in `status` will indicate where the logs actually got uplaoded
to.

Once we actually get to #107, and especially once we start supporting
endpoints other than GCS, we may find this isn't useful and remove it.

Fixes #146