Make pipelinerun reconciler call task resolution logic
Now when resolving a PipelineRun, we will resolve all of the resource
references as well. To do this, we make it so that PipelineRuns and
TaskRuns use the same types to bind resources. This has the bonus of
allowing users to provide resource paths in pipelineruns if they wanted
to.

This will allow us to use the same logic to validate resource bindings
for both pipelineruns and taskruns (for #213).