Use test builders and fix `providedBy` example 👷‍♀️
In https://github.com/knative/build-pipeline/pull/342 I added docs and
validation for the `providedBy` field, but I missed some feedback from
@vdemeester before the PR was merged, so I'm addressing it now!

The change to use the builders, even just for `PipelineResources`,
removes a surprising amount of boilerplate! It doesn't seem like I can
really use the builders for constructing the `state` object (slice of
`ResolvedPipelineRunTask`) unless we add builder for
`ResolvedPipelineRunTask`. That might be a good idea, however I'm
planning to iterate on some of this for #320 and #168 (maybe) so I'm
inclined to wait - open to other ideas tho!