Add more test cases and fix copy pasta 🍝
In the original `validate_resources_test`, we had a variable,
`validBuildSteps`, which wasn't actually being used and I carried that
over into the new tests.

In the validation itself, we weren't checking for nil values in the
`providedResources` map, which is a mistake I've made many times when
dealing with pointers. However in this case, we check for missing
resources at the beginning of the function and so should never see nil
values in the map. I added additional test cases to be sure this is
handled. And since making assumptions like this can make future
refactoring more difficult, I added the check anyway. (This makes me
suspicious that there is something I could be doing way better to
prevent this scenario but I'll leave that for a future iteration!)