Use `ResolvedTaskResources` builder in tests + fix tests 😅
Using the `ResolvedTaskResources` builder in these tests reduces the
boilerplate.

While making this change I also realized that the wrong resources were
being bound, which would still result in an error, but not the error
that the tests were actually trying to exercise. So I added some very
crude assertions on the message in the error to make sure we are
exercising the right errors in the future. I considered using custom
error types instead but asserting on the strings is simple and works for
now.

One issue I ran into with the builders - I wanted to use a
`PipelineTask` builder, but the function `PipelineTask` in the builder
package only works as a modifier for a `Pipeline`.